requirements.txt
google-genai==1.65.0
google-api-core==2.29.0
pillow==12.1.0
python-dotenv==1.2.1
websockets==16.0

# CUA Vision Agent dependencies
pyautogui>=0.9.54
pygetwindow>=0.0.9
numpy==2.4.3

# Browser Agent dependencies
browser-use==0.12.2
playwright==1.58.0
psutil==7.2.2

# TTS Integration (ElevenLabs)
requests==2.32.5
